Factor MatchingImplSet out of TypeChecker.
