+++ explorer/interpreter/type_checker.cpp
@@ -2122,186 +2122,6 @@ auto TypeChecker::RefineWitness(Nonnull<const Witness*> witness,
   }
 }
 
-// An impl match that we're currently performing.
-class TypeChecker::MatchingImplSet::Match {
- public:
-  explicit Match(Nonnull<MatchingImplSet*> parent,
-                 Nonnull<const ImplScope::Impl*> impl,
-                 Nonnull<const Value*> type, Nonnull<const Value*> inteface);
-  ~Match();
-
-  // Check that this match does not duplicate any prior one. Diagnose if it
-  // does.
-  auto Check(SourceLocation source_loc) -> ErrorOr<Success>;
-
- private:
-  class LeafCollector {
-   public:
-    LeafCollector(Match* match) : match_(match) {}
-
-    void Collect(const Value* value) {
-      value->Visit<void>(
-          [&](const auto* derived_value) { VisitValue(derived_value); });
-    }
-
-    void Collect(ValueKey key) {
-      ++match_->signature_[int(key)];
-    }
-
-   private:
-    // Most kinds of value don't contribute to the signature.
-    void VisitValue(const Value*) {}
-
-    void VisitValue(const TupleType* tuple) {
-      for (auto* elem : tuple->elements()) {
-        Collect(elem);
-      }
-    }
-
-    void VisitValue(const TypeType*) {
-      Collect(ValueKey::TypeType);
-    }
-
-    void VisitValue(const BoolType*) {
-      Collect(ValueKey::BoolType);
-    }
-
-    void VisitValue(const IntType*) {
-      Collect(ValueKey::IntType);
-    }
-
-    void VisitValue(const StringType*) {
-      Collect(ValueKey::StringType);
-    }
-
-    void VisitValue(const StaticArrayType* array) {
-      Collect(ValueKey::ArrayType);
-      Collect(&array->element_type());
-    }
-
-    void VisitValue(const PointerType* pointer) {
-      Collect(ValueKey::PointerType);
-      Collect(&pointer->pointee_type());
-    }
-
-    void VisitValue(const StructType* struct_type) {
-      for (auto [name, type] : struct_type->fields()) {
-        Collect(type);
-      }
-    }
-
-    void VisitValue(const NominalClassType* class_type) {
-      VisitDeclarationAndArgs(class_type->declaration(),
-                              class_type->bindings());
-    }
-
-    void VisitValue(const MixinPseudoType* mixin_type) {
-      VisitDeclarationAndArgs(mixin_type->declaration(),
-                              mixin_type->bindings());
-    }
-
-    void VisitValue(const InterfaceType* iface_type) {
-      VisitDeclarationAndArgs(iface_type->declaration(),
-                              iface_type->bindings());
-    }
-
-    void VisitValue(const NamedConstraintType* constraint_type) {
-      VisitDeclarationAndArgs(constraint_type->declaration(),
-                              constraint_type->bindings());
-    }
-
-    void VisitValue(const ChoiceType* choice_type) {
-      VisitDeclarationAndArgs(choice_type->declaration(),
-                              choice_type->bindings());
-    }
-
-    void VisitDeclarationAndArgs(const Declaration& declaration,
-                                 const Bindings& bindings) {
-      auto [it, added] = match_->parent_->declaration_keys_.insert(
-          {&declaration, match_->parent_->next_value_key_});
-      if (added) {
-        match_->parent_->next_value_key_ =
-            ValueKey(int(match_->parent_->next_value_key_) + 1);
-      }
-
-      Collect(it->second);
-      for (auto [key, value] : bindings.args()) {
-        Collect(value);
-      }
-    }
-
-   private:
-    Match* match_;
-  };
-
-  Nonnull<MatchingImplSet*> parent_;
-  Nonnull<const ImplScope::Impl*> impl_;
-  Nonnull<const Value*> type_;
-  Nonnull<const Value*> interface_;
-  llvm::DenseMap<int, int> signature_;
-};
-
-TypeChecker::MatchingImplSet::Match::Match(Nonnull<MatchingImplSet*> parent,
-                                           Nonnull<const ImplScope::Impl*> impl,
-                                           Nonnull<const Value*> type,
-                                           Nonnull<const Value*> interface)
-    : parent_(parent),
-      impl_(impl),
-      type_(type),
-      interface_(interface) {
-  // Build our signature.
-  LeafCollector collector(this);
-  collector.Collect(type);
-  collector.Collect(interface);
-
-  parent_->matches_.push_back(this);
-}
-
-TypeChecker::MatchingImplSet::Match::~Match() {
-  CARBON_CHECK(parent_->matches_.back() == this) << "match stack broken";
-  parent_->matches_.pop_back();
-}
-
-auto TypeChecker::MatchingImplSet::Match::Check(SourceLocation source_loc)
-    -> ErrorOr<Success> {
-  for (auto* match : parent_->matches_) {
-    if (match != this && match->impl_ == impl_) {
-      bool all_greater_or_equal = true;
-      bool any_greater = false;
-      for (auto [key, value] : signature_) {
-        int other_value = match->signature_.lookup(key);
-        if (value < other_value) {
-          all_greater_or_equal = false;
-          break;
-        }
-        if (value > other_value) {
-          any_greater = true;
-        }
-      }
-
-      if (all_greater_or_equal) {
-        if (any_greater) {
-          return ProgramError(source_loc)
-                 << "impl matching recursively performed a more complex match "
-                    "using the same impl\n"
-                 << "  outer match: " << *match->type_ << " as "
-                 << *match->interface_ << "\n"
-                 << "  inner match: " << *type_ << " as " << *interface_;
-        }
-        if (ValueEqual(match->type_, type_, std::nullopt) &&
-            ValueEqual(match->interface_, interface_, std::nullopt)) {
-          return ProgramError(source_loc)
-                 << "impl matching for " << *type_ << " as " << *interface_
-                 << " recursively performed a match for the same type and "
-                    "interface";
-        }
-      }
-    }
-  }
-
-  return Success();
-}
-
 auto TypeChecker::MatchImpl(const InterfaceType& iface,
                             Nonnull<const Value*> impl_type,
                             const ImplScope::Impl& impl,
